+++ releases/dogwood/3/bare/CHANGELOG.md
@@ -15,6 +15,7 @@ release.
 
 ### Changed
 
+- Make ORA2 configurable and use filesystem backend by default
 - Stop inheriting from MKTG_URL_LINK_MAP default setting
 
 ### Fixed